+++ packages/client/src/v2-events/layouts/workqueues/index.tsx
@@ -15,18 +15,14 @@ import { useIntl } from 'react-intl'
 import { useTypedParams } from 'react-router-typesafe-routes/dom'
 import { useSelector } from 'react-redux'
 import { AppBar, Button, Frame, Icon, Stack } from '@opencrvs/components'
-import { Plus } from '@opencrvs/components/src/icons'
-import {
-  canUserCreateEvent,
-  getAcceptedScopesByType
-} from '@opencrvs/commons/client'
+import { getAcceptedScopesByType } from '@opencrvs/commons/client'
 import { ROUTES } from '@client/v2-events/routes'
 import { ProfileMenu } from '@client/components/ProfileMenu'
 import { SearchToolbar } from '@client/v2-events/features/events/components/SearchToolbar'
-import { useEventFormNavigation } from '@client/v2-events/features/events/useEventFormNavigation'
+import { NewEventButton } from '@client/v2-events/features/events/NewEventButton'
 import { useAllWorkqueueConfigurations } from '@client/v2-events/features/events/useAllWorkqueueConfigurations'
 import { getScope } from '@client/profile/profileSelectors'
-import { useEventConfigurations } from '@client/v2-events/features/events/useEventConfiguration'
+import { useCreatableEventConfigurations } from '@client/v2-events/features/events/useEventConfiguration'
 import { emptyMessage } from '@client/v2-events/utils'
 import { constantsMessages } from '@client/i18n/messages/constants'
 import { Hamburger } from '../sidebar/Hamburger'
@@ -37,27 +33,13 @@ import { Sidebar } from '../sidebar/Sidebar'
  * @returns true if the user has the `record.create` scope for any event type, false otherwise.
  */
 export function useUserMayCreateEvents() {
-  const scopes = useSelector(getScope) ?? []
-  const eventConfigurations = useEventConfigurations()
-  return eventConfigurations.some(({ id }) => canUserCreateEvent(scopes, id))
+  return useCreatableEventConfigurations().length > 0
 }
 
 export function DesktopCenter() {
-  const { createNewDeclaration } = useEventFormNavigation()
-  const mayCreateEvents = useUserMayCreateEvents()
-
   return (
     <Stack gap={16}>
-      {mayCreateEvents && (
-        <Button
-          id="header-new-event"
-          type="iconPrimary"
-          onClick={createNewDeclaration}
-        >
-          <Plus />
-        </Button>
-      )}
-
+      <NewEventButton variant="header" />
       <SearchToolbar />
     </Stack>
   )